Remove unnecessary codegen test cases
Some of the cases were duplicated from the `Copy` tests but were not all
that relevant to a `Copy+Equatable` test.
