+++ crates/swift-integration-tests/src/opaque_type_attributes.rs
@@ -1,5 +1,4 @@
 mod already_declared;
 mod copy;
-mod copy_equatable;
 mod equatable;
 mod hashable;
